+++ src/client/cypress/e2e/detailPage/attachments.cy.ts
@@ -7,6 +7,7 @@ import {
   setTextInRow,
   uncheckAllVisibleRows,
   unCheckRowWithText,
+  verifyAnyRowContains,
   verifyPaginationText,
   verifyRowContains,
   verifyRowWithContentAlsoContains,
@@ -77,7 +78,7 @@ describe("Tests for 'Attachments' edit page.", () => {
 
       // check list of attachments
       verifyTableLength(1);
-      verifyRowContains("LOUDSPATULA.txt", 0);
+      verifyAnyRowContains("LOUDSPATULA.txt");
 
       // create file "IRATETRINITY.pdf" for input
       selectInputFile("IRATETRINITY.pdf", "application/pdf");
@@ -87,7 +88,7 @@ describe("Tests for 'Attachments' edit page.", () => {
       cy.wait(["@upload-files", "@getAllAttachments"]);
 
       verifyTableLength(2);
-      verifyRowContains("IRATETRINITY.pdf", 1);
+      verifyAnyRowContains("IRATETRINITY.pdf");
 
       // Upload and verify file "IRATETRINITY.pdf" for the second time but with different file name.
       selectInputFile("IRATETRINITY_2.pdf", "application/pdf");
@@ -95,14 +96,14 @@ describe("Tests for 'Attachments' edit page.", () => {
       cy.wait(["@upload-files", "@getAllAttachments"]);
 
       verifyTableLength(3);
-      verifyRowContains("IRATETRINITY_2.pdf", 2);
+      verifyAnyRowContains("IRATETRINITY_2.pdf");
 
       // Upload and verify file "WHITE   SPACE.pdf" to test file names with white spaces.
       selectInputFile("WHITE   SPACE.pdf", "application/pdf");
       cy.dataCy("addProfile-button").should("be.visible").click();
       cy.wait(["@upload-files", "@getAllAttachments"]);
       verifyTableLength(4);
-      verifyRowContains("WHITE___SPACE.pdf", 3);
+      verifyAnyRowContains("WHITE___SPACE.pdf");
 
       // Ensure files does not exist in download folder before download. If so, delete them.
       deleteDownloadedFile("IRATETRINITY_2.pdf");